Remove nextpoint()/makepoint() to draw as we go
This cleans up the code a lot, we used to fake a "yield" type mechanism.
makeapoint() would run the recursion until turtle.draw() was called to generate
_agraphic, at which point we would return. nextpoint() would draw the graphic
then call makeapoint() again. The glut display function repeatedly called nextpoint().
When things were slow, we used this to "interactively" draw the first points, then
single-buffer mode would just call nextpoint and add points as we got them.

Now all of this is just in draw().